leaf_area_web/FastSAM/requirements.txt
# Base-----------------------------------
matplotlib>=3.2.2
opencv-python>=4.6.0
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.64.0
fastapi
pandas>=1.1.4
seaborn>=0.11.0
pyserial
gradio==3.35.2
requests
git+https://github.com/openai/CLIP.git

# Ultralytics-----------------------------------
ultralytics == 8.0.120

